+++ src/Pim/Bundle/CatalogBundle/Entity/Family.php
@@ -391,6 +391,8 @@ class Family implements TranslatableInterface, ReferableInterface
         if (!isset($requirements[$requirementKey])) {
             $requirement->setFamily($this);
             $this->requirements->add($requirement);
+        } else {
+            $requirements[$requirementKey]->setRequired($requirement->isRequired());
         }
 
         return $this;
@@ -441,8 +443,8 @@ class Family implements TranslatableInterface, ReferableInterface
     {
         return sprintf(
             '%s_%s',
-            $requirement->getAttribute()->getCode(),
-            $requirement->getChannel()->getCode()
+            $requirement->getAttributeCode(),
+            $requirement->getChannelCode()
         );
     }