+++ src/Pim/Bundle/EnrichBundle/MassEditAction/Operation/Classify.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\MassEditAction;
+namespace Pim\Bundle\EnrichBundle\MassEditAction\Operation;
 
 use Pim\Bundle\CatalogBundle\Manager\CategoryManager;
+use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 
 /**
  * Batch operation to classify products
@@ -11,7 +12,7 @@ use Pim\Bundle\CatalogBundle\Manager\CategoryManager;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class Classify extends AbstractMassEditAction
+class Classify extends ProductMassEditOperation
 {
     /**
      * @var CategoryManager $categoryManager
@@ -74,15 +75,10 @@ class Classify extends AbstractMassEditAction
         return 'pim_enrich_mass_classify';
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function perform()
+    protected function doPerform(ProductInterface $product)
     {
-        foreach ($this->products as $product) {
-            foreach ($this->getCategories() as $category) {
-                $product->addCategory($category);
-            }
+        foreach ($this->categories as $category) {
+            $product->addCategory($category);
         }
     }
 }